refact project structure

+++ libs/task/source/taskSum.cpp
@@ -1,7 +1,7 @@
-#include "taskSum.h"
+#include "task/taskSum.h"
 
 float TaskSum::process( const std::vector< float >& data ) {
-    double result = 0;
+    float result = 0;
     for( uint32_t i = 0; i != data.size(); i++ ) {
         result += data[ i ];
     }
